initial kotlin codegen with gradle plugin, for now, for texture assets
right now it is ignorant of subdirs and hardcodes things, but it's a proof
of concept for now.
also add a test to better...test.

it'll be extended for fonts and sounds and shaders and arbitrary things
probably.

i chose to have it scan some dir, probably your source asset dir, for the
original assets before they hit the texture packer. this is so that it is
universal. my texture packing happens at startup, so a pre-build
scan of the atlas simply wouldn't ever compile (it'd be cyclical).

i'll have to see what people's use cases are for this sort of thing, as
i'm sure there will need to be changes to better fit them and add some
options/settings to the generator(s).